Added condition to delete user accounts

+++ app/controllers/registrations_controller.rb
@@ -59,15 +59,18 @@ class RegistrationsController < Devise::RegistrationsController
 
     # DELETE /resource
   def destroy
-    mentor = Mentor.find_by(user_id: current_user.id)
-    unless mentor.nil?
-      mentor.destroy
+    if Booking.find_by(mentee_id: current_user.id).nil?
+      mentor = Mentor.find_by(user_id: current_user.id)
+      unless mentor.nil?
+        mentor.destroy
+      end
+      resource.destroy
+      Devise.sign_out_all_scopes ? sign_out : sign_out(resource_name)
+      set_flash_message! :notice, :destroyed
+      yield resource if block_given?
+      respond_with_navigational(resource){ redirect_to after_sign_out_path_for(resource_name) }
+    else
+      redirect_to edit_user_registration_path, notice: 'Error: You can not delete your account because you still have booking sessions to attend to' 
     end
-    resource.destroy
-    Devise.sign_out_all_scopes ? sign_out : sign_out(resource_name)
-    set_flash_message! :notice, :destroyed
-    yield resource if block_given?
-    respond_with_navigational(resource){ redirect_to after_sign_out_path_for(resource_name) }
   end
-
 end
